refactor: separate `handleReportTemplateForm`
